Feature/disallow event removal

+++ app/Http/Controllers/ControllerDash.php
@@ -423,13 +423,19 @@ class ControllerDash extends Controller
     public function unsignupForEvent($id) {
         // Get the position request to be deleted
         $request = EventRegistration::find($id);
-
-        // Delete the request
-        $request->delete();
-
-        // Go back
-        return redirect()->back()->with('success', 'Your registration has been removed successfully.');
+        // Get date & time exactly 2 days from now
+        $two_days_from = new \DateTime('now +2 day');
+        // Check if event start time is less than 2 days from now
+        $within_two_days = $request->start_time < $two_days_from;
+        if (!$within_two_days) {
+            // Not within 2 days, delete registration
+            $request->delete();
+            return redirect()->back()->with('success', 'Your registration has been removed successfully.');
+        }
+        // Not within 2 days, send back with error and don't delete
+        return redirect()->back()->with('error', 'You cannot remove your registration within two days of an event.');
     }
+
 
     public function sceneryIndex(Request $request) {
         if($request->search == null) {
